+++ examples/captesting/src/main.rs
@@ -4,7 +4,7 @@ use nokhwa::utils::{CameraIndex, RequestedFormatType};
 use nokhwa::Camera;
 
 fn main() {
-    let index = CameraIndex::Index(50);
+    let index = CameraIndex::Index(0);
     let mut camera =
         Camera::open::<Mjpeg>(index, RequestedFormatType::AbsoluteHighestResolution).unwrap();
     println!("{}", camera.camera_format());